Add Flexibility to 1: /usr/NextSpace can be overridden ergonomically
Look, @trunkmaster is Our Hero for making
NextSpace and he can have it default to
anywhere-he-damn-well-pleases.

_But_

FreeBSD has the very nice and very sane religious
POV that userland applications live in
`/usr/local` not `/usr/`. So with a `DEST_DIR`, we
wound up with:

`/usr/local/usr/NextSpace`

as the install path. And that is an abomination.