7_build_libs-back.sh: Add FreeBSD compatibility
To get correct back-end ("art," as per existing
code) we need to pass additional data to the
Makefile.

I failed to understand the extent of the work
that's present in this vendor copy of back-art
(lib-back, for ART rendering).

ports(7) installs lib-back with Cairo by default.
This is clearly the future. It also, when used to
build Terminal.app (as I will do in subsequent
commits), will allow non-bitmap font rendering
(which my eyes need).

So the plan here is to install the FreeBSD version
from port(7) at /usr/ports/x11-toolkits/gnustep-back/ into
/usr/local/GNUstep. We'll link the Terminal.app
build against that.

For Nextspace otherwise, we'll build this back-art
version as well and install to
/usr/local/Nextspace

Other Fixes:

* Correct broken BACKVAL calc
* Install fonts
  These apps like Helvetica. Make sure you have
  something installed.
* Hack around bundle naming

  NextSpace wants to use art; thus, art-XXX.

  But it seems something wants this to be hard coded
  as back-XXX at app launch time (testing with
  OpenUp crashes).

fm squasher 7_build_libs-back.sh: Install 2 back-ends

Installs gnu-back version from port(7) at
/usr/ports/x11-toolkits/gnustep-back/ into
/usr/local/GNUstep. We'll link the Terminal.app
build against that.

For Nextspace otherwise, we link back-art version
as well and install to /usr/local/Nextspace

1. The port does:

    /usr/local/GNUstep/System/Library/Bundles/libgnustep-back-032.bundle

2. And NextSpace does:

    /usr/local/GNUstep/Local/Library/Bundles/libgnustep-art-032.bundle