.../SystemKit/FreeBSD/OSEFileSystem....m: Remove FIXME
Probably a mistake to remove, but this breaks in
compilation because of (I think) an API change.
It's also a relic from when @trunkmaster was
actively supporting *BSD, though and seems to be
bit-rotted.

OTOH, it's full of some pretty insightful
reasoning, **and** something similar exists for
Linux suggesting that some things Won't Work (tm)
without it.

On the other hand, I need to get to having an
application run that might need this so that I can
induce failures / test code paths.

Best way forward for now AFAICT.