TextEdit: Workaround broken NSFontPanel w/ env vars
Problem:

On FreeBSD, when launching TextEdit.app, I was
frustrated to always have a small font (default
from the GNUStep system). I need a bigger font for
comfort.

Step 1: Specify use of the Cairo rendering
back-end. NextStep prefers ART, but Cairo is the
future it's also a short path to OTF fonts.

Next I opened the Fonts (`#t`) menu item. This
launches `NSFontPanel`. I'm able to change fonts,
adjust sizes, etc. but the "Set" button _does
nothing_. I'm not sure if this is a FreeBSD thing,
but nothing persists.

Intended Path:

Clearly that wasn't meant to be the case.

> `NSFontManager` declares the programmatic
> interface to objects that manage font conversion
> in an application. `NSFontManager` is the center
> of activity for font conversion. `NSFontManager`
> accepts messages from font conversion
> user-interface objects such as the Font menu or
> the Font panel (see `NSFontPanel` for more
> details) and appropriately converts the current
> font in the selection by sending a `changeFont:`
> message up the responder chain

But for whatever reason, `Preferences` doesn't
receive this call. Consequently we don't change
the current document's font nor do we persist
changes to the app plist.

Solution:

Add a compiler flag that allows one, via 2 env
vars, to specify a font at app launch.