Adds plist showing how to get big, readable text
GNUstep documentation on 0 to usable is...not
friendly. Or reading it in a small font is so hard
I couldn't find out how to configure it to read it
in a big font.

Here's a plist showing what the
Preferences/Modules/Font widget winds up writing
and which makes (ahhhhh...) using NextSpace so
much easier on the eyes.