Preferences/Font: Call correctly named _setWMFont:forKey:
How did this ever work? I thought this was some
Obj-C slickness that I didn't or that the warnings
`clang(1)` was throwing didn't understand.

But, uh, it seems just this simple.

I'd change the system font and the NSFontPanel
would basically, it seems, idle. The `Set` button
would stay depressed. Walking it with logging, it
basically seemed to go nowhere. Looking at the
logs, I saw the warning and then asked Claude
about my syntax. It was like: "Uh, no, _ methods
are called with the _ in tact, pal."

So I changed it, and now everything works.