Bootup of NS is very slow as it repeatedly tries to test (~15 times) for several fonts. I really don't fully understand how font resolution is happening here and which name is used to do the lookups.
The cumulative effect of this is that we remove
the slow probing and we also find the fonts
although hacking FTFontEnumerator _and_ updating
the resolution path is probably unnecessary. One
of these approaches needs to get unwound, but I
need to try using the application more (with the
hack in place) so that I can see whether it's
stable.