7_build_libs-back.sh: Explain toggle in backend
Make it easy to change to Cairo. NS has religion
that we should use ART, but for those who need a
bigger or accommodating font, it's good to know
how.

However, let's not build it for users. Having to
handle a graphics back-end installation is taking
on more work than is implied by porting.  It
introduces some complexity that isn't required.